Factor out fileinfo to its own file
Move the file info struct to its own file. lib.rs is getting a bit big
